added timer

+++ src/Entities/Player.cpp
@@ -2,11 +2,11 @@
 // Created by Ettore Turini on 01/10/25.
 //
 
-#include "../includes/Player.h"
+#include "../../includes/Entities/Player.h"
 
 #include "SFML/System/Time.hpp"
 #include "SFML/Window/Keyboard.hpp"
-#include "../includes/Application.h"
+#include "../../includes/Core/Application.h"
 
 void Player::Update() {
 	sf::Vector2f alpha{0,0};
@@ -19,6 +19,6 @@ void Player::Update() {
 	if (sf::Keyboard::isKeyPressed(sf::Keyboard::Key::Down) || sf::Keyboard::isKeyPressed(sf::Keyboard::Key::S))
 		alpha.y += 1.f;
 	if (alpha.x != 0 || alpha.y != 0)
-		move(alpha.normalized()*Application::GetInstance()->GetDeltaT()->asSeconds()*move_speed);
+		move(alpha.normalized()*Application::GetInstance()->GetDeltaT()->asSeconds()*static_cast<float>(move_speed));
 
 }
